Reworked Raytracing to Use Camera Matrices
In the shader, GetRayDir now uses the normalized device coordinates
determined from gl_FragCoord, and some matricies provided by the
application to produce the appropriate ray direction.

If my math is correct, this should be reasonably close to if the
camera matrices were to be used in the conventional fashion to
rasterize similar primitives made from polygons.  I have not tested
this, but I think it would be worthwhile and reasonably easy to make
a "ground truth" comparison.

Also, the ray tracing pass does not currently write anything to the
depth buffer.

Something that might be interesting to explore later is having a ray
tracing pass per "object" in the scene.  This version would still rely
on screen space splats, but the area would be reduced to fit the rough
clip space bounds of the object.  This would also allow starting the
ray tracing much closer to the object.

I expect the performance benefitis of partitioning space for ray
tracing like this are 1) less time wasted plowing through negative space,
2) less of the scene has to be solved in a given pass, 3) somewhat
reduced program complexity.  On point 3, this would mean only have to
care about one world->object matrix, a smaller SDF tree to render, and
a smaller selection of relevant paint functions.  This probably also
would result in less branchy execution.

It would be nice to make the partitioned ray tracing thing optional
such that I could create some monolithic shaders that can be translated
to shadertoy moderately easily.  However, the partitioned space thing
could probably be adapted to work for shadertoy as well.

Partitioned whachamacalit or not, I think what I want to do nextish is
to try to build something interesting with this to work out how I want
to change this towards producing some kind of rendering engine I could
make a game with.