Start of Drawing Objects in Separate Passes
This change moves the various world->local space transforms out of
the shader, and introduces a mechanism for the application to switch
out which object should be ray traced.

As a result, now we are doing 3 full screen passes instead of just one
with no special mechanism to reduce the search area.

This change also corrects a diabolical NaN poisoning bug in the RayTrace
function where infinite test points would be produced in the negative space
surrounding the shapes, which would then turn into NaNs after multiplying
against a typical transformation matrix.  So long as nothing attempted to
clamp the value, the rendered image would actually look correct.

The Box signed distance function is not one of these cases, so if a box
object were to be drawn, the NaNs in the negative space around it would be
clamped to 0, resulting in the entire screen being washed with whatever
paint function that object used.

In the other news, the shader now writes out the appropriate depth value,
and depth testing is now active.

The obvious thing to do next is to limit the tracing area per object based
on the bounding box, which should be pretty easy now.