World Normals and Simple Lighting
This change adds the functions WorldNormalViaGradient and
WorldNormalViaDerivatives, which are used for generating approximate worldspace
normals.  The later is much cheaper, but produces noticable artifacts.

I think in the future it would be good to generate normals from the shape
functions, and only fall back to approximate normals when a shape operator or
something obliterates the more precise value.

This change also improves the accuracy of the RayTrace function by doing an
extra refinement step after an acceptable SDF is found.  This also does away
with some of the limiting mechanisms I put in place earlier, since the
DepthRange check seems to be sufficient to prevent rays from flying off to
infinity.

I think there is something wrong with how my coordinates are setup or such.
Things are flipped from where I expect them to be, so I guess there is
another math audit in the future as well.