Fixed Coordinate System
This change sets the screen space origin back to the lower left
corner.  This was causing the scene to be flipped, since GLM assumes
the origin is upper left.

I'm sure there is probably a way to make GLM do what I wanted, but
this is fine, too.

This change also revises the test scene etc to make the coordinate
system correctness or lack theroff more readily apparent.