Reworked Tangerine Function, Added Lime
This change has the "tangerine" paint function produce the apparent
internal structure of the citrus fruits, rather than require this be
established during ray marching.

I realized that for something like an orange, the only thing that really
mattered during ray tracing was a sphere, or whatever, and that further
SDF stuff for internal structures could be moved to the paint function
just fine.  This also allows for some simplification, since the only
concern is producing the correct output color.  This approach may also be
a little faster than what I had before, but the profiling numbers are
noisy and the scene is not set up as a stress test, so it is kind of hard
to say.

This change also adds a lime to the scene.