Draw All of Everything at Once Via Instancing
This only works in as much as there is only one shader program for
the entire scene, as is the case right now.  Since each object is
drawn with the same sort of screen space quad, this works out nicely.

This also changes it so all of the scene information is consolidated
in one shader storage buffer.  This is currently only accessible via
the vertex shader, but this may be handy to later provide some sort
of volume hierarchy for doing additional screen-wide ray tracing for
things like shadows.