RayMarching Improvements, Started Implementing Cube Tracing Path
This change restructures the raymarching shader so that there is a
(broken) cube tracing path that may be used for cube objects instead
of the normal raymarching path.  The cube tracing path is currently
disabled, so everything should still be going down the ray marching
path.

This change also brings the main improvement of doing the ray marching
in an object's local space instead of world space.  This prevents the
need to multiply the WorldToLocal transform every step.  This greatly
cuts down on the ALU load.  For the current test scene of about 10004
objects, this cuts the draw time down by about a millisecond on a
RTX 2080 Ti.