Quick and Dirty "CubeTrace" Implementation
This version determines the six different ray distances which might be
valid box intersections.

Rather than determining which one is correct, it just tests all of them,
and whichever one is the lowest ray distance with a solid sdf value.

This means that CubeTraceSceneSDF always gets called six times.  This
is not optimal, but this beats the pants of the worst case of calling
SceneSDF like 65+ times in RayMarch.

Perfwise, enabling CubeTrace does better than leaving it disabled, so
it is enabled now.

There is a lot of room for improvement here still for sure.

An alternative I'm considering is have CubeTrace find the distance
that hits the first side of the AABB and then ray march from there.
This might actually just be ideal approach for all ray marching for
objects where an AABB is the most reasonable bounding shape.