CubeTrace Sorts to Reduce CubeTraceSceneSDF Calls
RayDists is now sorted from lowest to highest, and then we just test
them all in order and stop at the one that matches a surface first.

I think this might be a little faster, but it is hard to say.
I intend to replace this with a more conventional approach to AABB/Ray
intersection later.