RayMarch Variant and CubeTrace Bug Fix
This changes CubeTrace so that after we sort RayDists, we iterate
starting from index 0, not 1.  We also stop after the third, because
the last three will be behind the shape.

As for RayMarch, this change adds a variation where we start basically
the same as CubeTrace to determine the ray start position, but then
from there we just ray march the rest.  This also adjusts the maximum
distance check to be the last value in RayDists.

The case where the ray origin is within the shape is handled wrong -
it should just trace outward from there.  That'll have to be addressed
later if I choose to keep this approach.

The perf difference between the two RayMarch variants seems to be a
wash right now, but the number of relevant objects in the test scene
is low and their bounding boxes are not necessarily ideal.

Leaving the start-with-cube-tracing-but-then-ray-march variant on for
now.