Sketched Out A Different Ray Marching Approach
The file "scratch.py" contains simple python implementations for two
different ray marching approaches.

The function "sphere_march" is the old approach I've been using, where
you start at the beginning of the ray, and loop reading the scene sdf to
determine how far to advance each iteration.

The function "search" is my new approach that I'm considering using.
In this approach, it starts by reading the scene sdf at the middle of
ray, and then iterates by adding more samples to fill covearge along
the ray.  This approach has the advantage that less space is redundantly
"scanned" by scene sdf reads than the sphere marching approach.

For dead on hits, performance is about the same between the two, but in
cases like acute angles or running parallel to an object, the "search"
approach performs substantially better, in terms of number of scene sdf
samples.

So, that is great.  Now to figure out how to implement this in a way
that is at all sensible for a shader.