GLSL Implementation of the New Ray Marching Algorithm
With cube tracing on, this version appears to perform marginally
better than my current sphere tracing implementations.  However,
if cube tracing is disabled, then this version performs worse.

I think this is consistent with my experiments so far with the
python implementation of this algorithm.  There are also a few
obvious areas for improvement on this implementation that are
highlighted as code comments.

I believe that this will get better average performance for scenes
with lots of complex objects, and for scenes with mostly simple
objects like my current test scenes should use cube tracing etc for
those objects (which is also true for my sphere tracing
implementation).