Split Shape Functions into Color and Hull Variants to Speed Up Ray Marching
This unfortunately means that objects currently have to be described twice,
but this also means lowering register pressure in the RayMarch functions.
This makes quite a difference in overall speed, so I think the verboseness
is very munch an acceptable tradeoff.

Currently RAYMETHOD_CUBE_ELIMINATE is fastest, so that one is default for now.