Exposed Elapsed Time to Shader, Playing Around
This removes the camera info block and folds it into the view info block.
The elapsed time is written into the previously unused w component of
CameraOrigin.

I also played around with the Tangerine* shape functions a bit, though the
changes are dummied out.  I also realized that only the *Hull functions had to
have the correct topology, so this change simplifies some of the *Color
functions.