Fixed Culling Ha Ha Wow
The culling code only checked the clip space X/Y bounds, but did not actually
check if the object was behind the camera or not.  Whoops!

This makes *quite* a difference in perf, especially on the "random terrain"
scene, bringing it down from about 30ms for the instanced draw to about 2ms
on the RTX 2080 Ti!