Split Rendering into a Depth Pass and Color Pass
The depth pass does the ray tracing as before, but it only ever calls SceneHull
and outputs a depth value and an object id.

The Color pass does not do any ray tracing.  The color pass is a single
full-screen draw call, and for the most part is just everything needed to
either draw the clear color or call SceneColor, based on the data from the
depth pass.