Rudimentry Resolution Scaling
When active, this scales the depth pass down by an arbitrary amount.  The color
pass then samples the depth and object ID buffers w/ the nearest filter.

The color pass generates its world position via the sampled depth and the
calculated ray dir.  This will result in a different world position than was
generated from the depth pass.

The problem with this approach is that the color pass's world position will
often be far enough outside the surface of the object that the paint function
will bug out.  This also introduces a ton of aliasing.

I think to make this workable, there needs to be an upsample pass between the
depth pass and the color pass.  Such a pass would be responsible for also
refining the depth value when possible.