Improved Profiling
This change now tracks the depth and color passes w/ separate timing queries,
changes how stat averages are tracked, restores CPU frame stats, and prints the
log out to stdout in one write so that it synchronizes better with terminals.

Averages use a ring buffer as before, but now they are just rolling averages.
The GPU stats are on a 100 frame average, whereas the CPU stats use a 50 frame
average.  The ring buffer sizes are determined unscientifically for a balance
between readability and responsiveness.

Since the log is a string stream now, this could also make it easier to overlay
the stats on screen in the future.