Sun Shadows!
This change adds some basic ray traced shadow capabilities.

Objects can be marked as shadow casters.  Currently all objects receive
shadows, however they cannot self-shadow.

To facilitate this, a buffer was added to track the coverage information per
object.  When the color pass is ran, the ObjectId of the current pixel is used
to pull up the occluder candidates for the object, which are then ray traced
towards the light until a hit is found.

This should probably be pulled out of the color pass and into its own pass
later.

Currently there can be at most four occluder candidates per pixel.  This works
ok if you're only expecting characters and a few other things to cast shadows
and the sun is almost directly overhead, but this falls appart with the forst
scene when the sun is low.