Fixed Text Rendering, Integer Buffer Usage
The hack to use a floating point buffer to get text rendering working
produced plausible results on nvidia, but corrupted results on Intel.

This change corrects the texture upload for logging to use
GL_RED_INTEGER instead of GL_RED, and sets the format back to GL_R8I.

This also changes the ID buffer back to using an integer value, which
I changed for some reason earlier.  This works as expected on both
Intel and Nvidia.