Prototyping the "Gloom Pass"
Working on mixing rendering arbitrary geometry against the raymarched depth
buffer from the depth pass.

I don't think the depth values are quite compatible, but they're close enough
that this might be fine for bounding volumes, at least for the near term.  I'd
like to get them to line up perfectly.

This also comes with the discovery that I wasn't using zero-to-one clipping
correctly, which probably comes at a precision loss.  GLM's infinitePerspective
appears to be incompatible with zero-to-one clipping, so I've switched this
over to negative-one-to-one clipping.  See glClipControl for more details.

Interestingly, glm's "perspective" and "perspectiveFov" functions have variants
for zero-to-one if GLM_FORCE_DEPTH_ZERO_TO_ONE is set.

So, I probably want to either:

 1. Switch to "perspective" or "perspectiveFov"?
 2. Dig through papers and try to find an adequate replacement for
    "infinitePerspective"?
 3. Stick with negative-one-to-one?
 4. ???

The other problem here is I suspect that with the perspective matrix, depth is
constant along a given z-plane, as apposed to how I'm using the depth buffer w/
the depth and color passes, wherein depth is the distance from the point to the
origin.

For the sake of being able to do early-z elimination for the gloom passes, the
depth and color passes should be updated so that everything is using consistent
math.

I'm also not totally sure that the ray dir generation scheme I'm using is
actually correct, but it seems close enough for now.

So, the followups to this change are probably:

 1. Update cube.vs.glsl so we can draw arbitrary boxes!
 2. Generate shadow volumes for occluders.
 3. Instanced draw for all of the shadow volumes for the gloom pass.
 4. Implement the fs logic for the gloom pass.
 5. Do something clever with the results in color.fs.glsl.

It is tempting to instead do something like this though:

 1. Instanced draw for all of the shadow casters for the gloom pass.
 2. Implement the fs logic for the gloom pass.
 3. Do something clever with the results in color.fs.glsl.
 4. Return to the boundary volume stuff later.