Gloom Pass
This change is a followup to commit d4739b59a698c7f84ec70ce58a8b0a8e5f0c6951.

This change sets aside the mixed raster / raymarch rendering stuff I started on
in favor of building out the new shadow casting system.

This version does a full screen quad per shadow caster for the gloom pass,
wherein the shadow coverage is determined for each pixel and drawn out into a
buffer.  The results are combined together using the GL_MIN blend equation.

For now, the color pass reads the result from the gloom buffer to determine the
light intensity mask value instead of having to do its own raymarching.

The advantages to this approach are:
 1. Very little per-frame work is needed CPU side to support it.
 2. A given pixel can be covered by any number of occluders instead of just four.
 3. This generalizes well to adding multiple light sources later on.
 4. There are several very obvious optimizations that can be applied later on.

The downside to the current implementation is that everything shadow caster is
a full screen pass, so, it is burning a lot of GPU time.