Light Transmission Experiment
When ENABLE_LIGHT_TRANSMISSION is enabled, the Gloom pass will determine how
much light from the light source is transmitted to the occluded surface.

When an occluder is hit, an incremental search through the occluder determines
the transmission rate.

The numbers and approach are unscientific.

I'm thinking of changing this up so that after the first occluder hit is found,
it runs the occlusion trace again backwards to find the last hit.  Then it would
take samples (probably along a discrepancy sequence) to determine the density
and so on.  This probably will work by attenuating per sample, and if the last
sample was air gap, use an SDF for both samples to guess as to what the distance
to the solid sample is etc.  Air gap regions should maybe have the option of
deattenuating based on some ambient light factor?