Better Gloom Pass Light Transmission Prototype
This commit changes how gloom pass light transmission works to have more robust
attenuation.

When light transmission is enabled, transmissive shapes will call
"TransmissiveSearch" instead of "OcclusionRayMarch" in the gloom pass.  This
will then raymarch 20 iterations to match both the first and last hit of the
shape, or as close as it can get.  This is used to determine the search area,
which is then iterated upon over a fixed interval.  The iterations sample which
transmission "bucket" a given point belongs to, to determine the overall
density of the different materials that might make up a transmissive object.
Afte that, the lighting contributions are determined and combined.

Non-transmissive shapes effectively have the same occlusion path as would be
used if transmissive lighting were disabled.

This hard-codes a bunch of stuff specific to the orange, some followup work
is needed to generalize this so that it can be used with other shapes.