Added "VINE_MODE"
When VINE_MODE is set true, the program will output a six second video instead
of rendering the scene continuously.  The raw output frames will be in the
"frames" folder, which also contains generate_mp4.sh, which just wraps a ffmpeg
invocation to produce a video that is suitable for upload to Twitter or Byte.