Fixed Artifacts When Light Rays are Parallel to Surfaces
The artifact will still be present in the gloom buffer, but it should not
affect the final lighting computation.

The lighting computation now uses the illumination as a blend factor between
the material's provided base color, and a calculated shadow color.  This
shadow color could be later determined using some sort of ambient component,
and/or be specified by the material as well.