Resizable Buffers, Removed Dead Code
This updates the Buffer struct so that the underlying gl buffer is deleted when
the struct is destructed or when Upload is called with a new size.

The Buffer struct also now has a constructor so that a gl object label can be
optionally specified.

I noticed that the BufferPool struct was not used by anything, so I've removed
it.

This change also updates the behavior of the demo so that the buffers for
visible objects and shadow casters are tightly sized rather than being fixed to
the total number of objects in the scene.  Because of this, it is possible now
to dynamically add and remove objects from the scene, eg for level loading or
spawning gameplay entities.