Reorganizing SDFExperiment::Render Into Separate Functions
This change breaks apart SDFExperiment::Render into several smaller functions.
This is to make it easier to reorganize rendering passes, as well as to make it
easier to read and edit.

This change also renames the framebuffer and texture handles used by vine mode
to "FinalPass" and "FinalBuffer".

This change adds a framebuffer and two texture handles ("ColorPass" and
"ColorBuffers"), which will be used by a followup commit.