Moved Lighting Into Own Pass
This change moves the color pass before the gloom pass, and adds a "light"
pass.

The color pass now outputs the world normal, the material color, and whether or
not the material is transmissive.

The new lighting pass reads the data from the color pass and the gloom pass to
produce the final lighting.

From here, I think it should be relatively easy to support multiple light
sources: the gloom pass and the light pass would run once per light, such that
the light pass reads from it's corresponding gloom pass and accumulates energy
into a singular buffer.

With multiple lights, a tone mapping pass would be needed to assemble the final
image.

One approach I'm considering is that the lighting pass would only read the
parameters needed for determining the light propogation (eg, if I ever added
anything like "roughness").  After tone mapping the accumulated light, the
result might be then used as an alpha value for lerping between the base color
and shadow color of the material.  There is also a case where I would want to
be able to override the shadow color so that objects can determine the color
of their shadows', eg to emphasize an object.  I haven't decided how that is
meant to fit into this though.

The advantage of this strange approach would be greater artistic control over
the use of color and contrast in the rendered image.  If I decide to run with
it, it may be that material properties should be thrown into something like a
color pallet, and the base color "gbuffer" could then be replaced with an
index or something.

I'm not sure at all how anti-aliasing should fit into this at the moment.

Oh, that reminds me, this change breaks AA, whoops.