Tiled Gloom Pass
This change introduces a variation for the gloom pass where rather than draw a
single full screen quad, the screen is divided into 8x8 quads.  The vertex
shader then does a simple cube trace test for the entire quad, to see if it
could be occluded.  The depth range is used to produce a radius for the tile,
which the occluder's AABB is increased by so that the cube trace will work.

This might only work for sun rays, I haven't really thought it through as far
as point lights or spots.

The radius is also wrong, albiet really close.  I think it needs to actually
be the diagonal of the froxel, not the depth distance.  Good enough for now,
though.

Anyways, I think this could be a good place to try using a mesh shader.  The
conversion should be pretty simple, and the gains should scale up with the
resolution.