Correct Gloom Tiling Behavior; Point Tiles Experiment
This change corrects two problems in the gloom vertex and pixel shaders:

 1. The gloom fragment shader's early out condition for sky tiles was
    previously removed.

 2. The gloom vertex shader had the wrong math for mapping tiles to screen
    space, resulting in overdraw and in some cases inaccurate depth ranges.

Both of these issues have been corrected now.

This change also introduces an alternative mode of drawing tiles using point
primitives.  Somehow it is *slower* though, so it is disabled for now, and
will probably be removed eventually if it doesn't prove to be useful on
other hardware.