Reduce Exploding Verticies from Smooth Union
Still needs work.  The stray verts are still from when a smooth union operand
expands some of its verts to the other side of the other operand.  Or really
the middle, etc.

The SceneFn function needs some work still I think.  Or maybe the function needs
to be paired with culling that checks the smooth union verts against the ideal
version.