Edge Slide in fine.gs.glsl to Decouple From Smooth Mesh Refinement
This fixes the weird spikes along the edge.  Now all that realy remains is
to tweak the tessellation and smooth mesh reflow logic so that the results look
nice, figure out what to do with the final draw FS, better culling, put in some
timing queries, and write up a nice post about how it all works.